fix #22, handle "else if" better
use regular expression to match tokens instead of splitting words, in
this way "else if" can be matched.

This is painful to consider every scenario and match it by regex.
There still are plenty of corner cases where the script indents wrongly.
The better way is to understand the semantics,
which might take a while to be done.